codes refactor
