+++ controller/role_api.py
@@ -4,7 +4,7 @@ from flask import request, Response
 from services import RoleService
 from utils.common import json_response
 from controller import api_blueprint as controller
-from utils.test import validRoleObject, invalidObjectMessage
+from utils.validatepayload import validRoleObject, invalidObjectMessage
 
 
 @controller.route('/role', methods=['GET'])
@@ -22,6 +22,6 @@ def create_role():
     request_data = request.get_json()
     if (validRoleObject(request_data)):
         RoleService.create(request_data)
-        return Response('', 200, mimetype='application/json')
+        return Response(json.dumps(request_data), 200, mimetype='application/json')
     else:
         return Response(json.dumps(invalidObjectMessage), 401, mimetype='application/json')
