+++ controller/users_api.py
@@ -4,7 +4,7 @@ from flask import request, Response
 from services import UserService
 from utils.common import json_response
 from controller import api_blueprint as controller
-from utils.test import validUserObject, invalidObjectMessage
+from utils.validatepayload import validUserObject, invalidObjectMessage
 
 
 @controller.route('/users', methods=['GET'])
@@ -21,7 +21,7 @@ def create_users():
 
     if (validUserObject(request_data)):
         UserService.create(request_data)
-        return Response('', 200, mimetype='application/json')
+        return Response(json.dumps(request_data), 200, mimetype='application/json')
     else:
         return Response(json.dumps(invalidObjectMessage))
 
@@ -31,7 +31,7 @@ def update_users(user_id):
 
     if(validUserObject(request_data)):
         UserService.update(user_id,request_data)
-        return Response('',200,mimetype='application/json')
+        return Response(json.dumps(request_data),200,mimetype='application/json')
     else:
         return Response(json.dumps(invalidObjectMessage))
 
@@ -39,6 +39,6 @@ def update_users(user_id):
 def delete_users(user_id):
     if(UserService.delete(user_id)):
         UserService.delete(user_id)
-        return Response('deleted', 200, mimetype='application/json')
+        return Response('', 200, mimetype='application/json')
     else:
         return Response('Unable to perform delete operation',401,mimetype='application/json')
