+++ models/sc_class.py
@@ -1,4 +1,4 @@
-from models.base_model import BaseModel, db,DbModel
+from models.base_model import BaseModel, db,DbModel,session
 
 
 class SchoolClass(BaseModel,DbModel):
@@ -14,3 +14,17 @@ class SchoolClass(BaseModel,DbModel):
             'grade': self.grade,
             'section': self.section
         }
+
+    @classmethod
+    def create(cls, requested_data):
+        new_class = SchoolClass(grade=requested_data['grade'], section=requested_data['section'])
+        session.add(new_class)
+        session.commit()
+
+    @classmethod
+    def update(cls, _class_id, requested_data):
+        update_class = SchoolClass.query.filter_by(class_id=_class_id).first()
+        update_class.grade = requested_data['grade']
+        update_class.section = requested_data['section']
+        session.add(update_class)
+        session.commit()
