+++ services/base_service.py
@@ -14,3 +14,12 @@ class BaseService:
     @classmethod
     def delete(cls,id):
         return cls.model.delete(id)
+
+    @classmethod
+    def create(cls,requested_data):
+        cls.model.create(requested_data)
+
+    @classmethod
+    def update(cls,id,requested_data):
+        cls.model.update(id,requested_data)
+
