+++ services/role_service.py
@@ -5,8 +5,3 @@ from services.base_service import BaseService
 class RoleService(BaseService):
     model = Role
 
-    @classmethod
-    def create(cls, request_data):
-        new_role = Role(type=request_data["type"])
-        session.add(new_role)
-        session.commit()
