+++ services/student_service.py
@@ -6,25 +6,6 @@ class StudentService(BaseService):
     model = Student
 
     @classmethod
-    def create(cls, request_data):
-        new_student = Student(name=request_data["name"], father_name=request_data["father_name"],
-                              mother_name=request_data["mother_name"],
-                              contact_number=request_data["contact_number"], class_id=request_data["class_id"])
-        session.add(new_student)
-        session.commit()
-
-    @classmethod
-    def update(cls, _student_id, _requested_date):
-        updated_student = Student.query.filter_by(student_id=_student_id).first()
-        updated_student.name = _requested_date['name']
-        updated_student.father_name = _requested_date['father_name']
-        updated_student.mother_name = _requested_date['mother_name']
-        updated_student.contact_number = _requested_date['contact_number']
-        updated_student.class_id = _requested_date['class_id']
-        session.add(updated_student)
-        session.commit
-
-    @classmethod
     def fetch_students_by_class_id(cls, _class_id):
         return Student.query.filter_by(class_id=_class_id)
