+++ services/subject_service.py
@@ -7,20 +7,6 @@ class SubjectService(BaseService):
     model = Subject
 
     @classmethod
-    def create(cls, request_data):
-        new_subject = Subject(name=request_data["name"], class_id=request_data["class_id"])
-        session.add(new_subject)
-        session.commit()
-
-    @classmethod
-    def update(cls, _subject_id, _requested_date):
-        updated_subject = Subject.query.filter_by(subject_id=_subject_id).first()
-        updated_subject.name = _requested_date['name']
-        updated_subject.class_id = _requested_date['class_id']
-        session.add(updated_subject)
-        session.commit()
-
-    @classmethod
     def fetch_subject_by_class_id(cls, _class_id):
         return Subject.query.filter_by(class_id=_class_id)
